car-rental-backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.1
email-validator==2.2.0
pydantic-settings==2.9.1
python-dotenv==1.1.0
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
alembic==1.15.2
greenlet==3.2.2
motor==3.7.0
redis[hiredis]==6.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
types-passlib==1.7.7.20240819
python-jose[cryptography]==3.3.0
types-python-jose==3.3.4.20240106
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
